Enemy AI update
Enemy AI seems to be working without bugs. allowing the enemies to
select a skill to use (based on several different types) and a target to
select (out of the allied set and enemy party, based on logical actions
eg. finish off the low target or heal the hurt ally)